requirements.txt
asgiref==3.5.0
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.2
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-spectacular==0.21.2
idna==3.3
importlib-metadata==4.10.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.4.0
Markdown==3.3.6
MarkupSafe==2.0.1
oauthlib==3.2.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyrsistent==0.18.1
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
zipp==3.7.0
django-neomodel==0.0.7
neo4j-driver==4.3.1
neomodel==4.0.4
whitenoise==3.3.1
black==22.1.0
flake8==4.0.1
django-webtest==1.9.9
Faker==12.3.0
factory-boy==3.2.1
django-discover-runner==1.0
coverage==6.3.1
Pillow==9.0.1